+++ src/main/java/com/aacademy/moonlight/repository/hotel/RoomFacilityRepository.java
@@ -5,5 +5,5 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface RoomFacilityRepository extends JpaRepository<RoomFacility, Long> {
+public interface RoomFacilityRepository extends JpaRepository<RoomFacility, Integer> {
 }
